tests/wpt/web-platform-tests/keyboard-lock/navigator-cancelKeyboardLock.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_equals(navigator.cancelKeyboardLock(),
                undefined);
}, 'Keyboard Lock cancelKeyboardLock');

</script>
